requirements.txt
langchain==0.1.14
langgraph==0.0.33
langchain-openai==0.1.1
langchain-pinecone==0.0.3
pinecone-client==3.0.2
openai==1.14.2
beautifulsoup4==4.12.3
tqdm

